+++ dispatch/workspace.go
@@ -259,3 +259,19 @@ func ensureWorkspace() string {
 
 	return dispatchUID
 }
+
+func EnsureDependencies(event Event) Event {
+	fmt.Print("\nEnsuring dependencies:\n")
+
+	event.User = ensureWorkspace()
+
+	clientConfig := awsClientConfig()
+
+	testAWSCreds(*clientConfig)
+
+	event.Bucket = ensureS3Bucket(*clientConfig, event.User)
+
+	printExistingClusters(event.Bucket)
+
+	return event
+}
